"refactor: create separate file for attribute class"
